[B2BORG-76]
"Modify orgs functionality for adding vs editing user accounts"
- the structure of the components has been improved;
- Added 2 mutations in order to follow the mutations from storefront permissions;
- Remove a couple of conditionals nested;
- Remove unnecessary variables, some changes to inline returns;
